Update maps implementation proposal
