--- ogcapi-draft/ogcapi-maps/build.gradle
@@ -1,3 +1,0 @@
-dependencies {
-    provided project(":ogcapi-tiles")
-}
