+++ ogcapi-draft/ogcapi-maps/src/main/java/de/ii/ldproxy/ogcapi/maps/domain/MapTilesConfiguration.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package de.ii.ldproxy.ogcapi.maps;
+package de.ii.ldproxy.ogcapi.maps.domain;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import de.ii.ldproxy.ogcapi.domain.ExtensionConfiguration;
