+++ ogcapi-draft/ogcapi-tiles/src/main/java/de/ii/ldproxy/ogcapi/tiles/infra/EndpointMapTileMultiCollection.java
@@ -8,28 +8,19 @@
 package de.ii.ldproxy.ogcapi.tiles.infra;
 
 import com.google.common.collect.ImmutableList;
-import de.ii.ldproxy.ogcapi.collections.domain.ImmutableOgcApiResourceData;
 import de.ii.ldproxy.ogcapi.domain.ApiEndpointDefinition;
-import de.ii.ldproxy.ogcapi.domain.ApiOperation;
 import de.ii.ldproxy.ogcapi.domain.ApiRequestContext;
-import de.ii.ldproxy.ogcapi.domain.ConformanceClass;
 import de.ii.ldproxy.ogcapi.domain.ExtensionConfiguration;
 import de.ii.ldproxy.ogcapi.domain.ExtensionRegistry;
-import de.ii.ldproxy.ogcapi.domain.FormatExtension;
-import de.ii.ldproxy.ogcapi.domain.HttpMethods;
-import de.ii.ldproxy.ogcapi.domain.ImmutableApiEndpointDefinition;
 import de.ii.ldproxy.ogcapi.domain.OgcApi;
 import de.ii.ldproxy.ogcapi.domain.OgcApiDataV2;
-import de.ii.ldproxy.ogcapi.domain.OgcApiPathParameter;
-import de.ii.ldproxy.ogcapi.domain.OgcApiQueryParameter;
 import de.ii.ldproxy.ogcapi.features.core.domain.FeaturesCoreProviders;
 import de.ii.ldproxy.ogcapi.tiles.domain.StaticTileProviderStore;
 import de.ii.ldproxy.ogcapi.tiles.domain.TileCache;
-import de.ii.ldproxy.ogcapi.tiles.domain.TileFormatExtension;
 import de.ii.ldproxy.ogcapi.tiles.domain.TilesConfiguration;
 import de.ii.ldproxy.ogcapi.tiles.domain.TilesQueriesHandler;
-import de.ii.ldproxy.ogcapi.tiles.domain.tileMatrixSet.TileMatrixSetRepository;
 import de.ii.ldproxy.ogcapi.tiles.domain.tileMatrixSet.TileMatrixSetLimitsGenerator;
+import de.ii.ldproxy.ogcapi.tiles.domain.tileMatrixSet.TileMatrixSetRepository;
 import de.ii.xtraplatform.auth.domain.User;
 import de.ii.xtraplatform.crs.domain.CrsTransformationException;
 import de.ii.xtraplatform.crs.domain.CrsTransformerFactory;
@@ -54,34 +45,29 @@ import java.util.List;
 import java.util.Optional;
 
 /**
- * Handle responses under '/tiles/{tileMatrixSetId}/{tileMatrix}/{tileRow}/{tileCol}'.
+ * Handle responses under '/map/tiles/{tileMatrixSetId}/{tileMatrix}/{tileRow}/{tileCol}'.
  */
 @Component
 @Provides
 @Instantiate
-public class EndpointTileMultiCollection extends AbstractEndpointTileMultiCollection implements ConformanceClass {
+public class EndpointMapTileMultiCollection extends AbstractEndpointTileMultiCollection {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(EndpointTileMultiCollection.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(EndpointMapTileMultiCollection.class);
 
     private static final List<String> TAGS = ImmutableList.of("Access multi-layer tiles");
 
     private final FeaturesCoreProviders providers;
 
-    EndpointTileMultiCollection(@Requires FeaturesCoreProviders providers,
-                                @Requires ExtensionRegistry extensionRegistry,
-                                @Requires TilesQueriesHandler queryHandler,
-                                @Requires CrsTransformerFactory crsTransformerFactory,
-                                @Requires TileMatrixSetLimitsGenerator limitsGenerator,
-                                @Requires TileCache cache,
-                                @Requires StaticTileProviderStore staticTileProviderStore,
-                                @Requires TileMatrixSetRepository tileMatrixSetRepository) {
+    EndpointMapTileMultiCollection(@Requires FeaturesCoreProviders providers,
+                                   @Requires ExtensionRegistry extensionRegistry,
+                                   @Requires TilesQueriesHandler queryHandler,
+                                   @Requires CrsTransformerFactory crsTransformerFactory,
+                                   @Requires TileMatrixSetLimitsGenerator limitsGenerator,
+                                   @Requires TileCache cache,
+                                   @Requires StaticTileProviderStore staticTileProviderStore,
+                                   @Requires TileMatrixSetRepository tileMatrixSetRepository) {
         super(providers, extensionRegistry, queryHandler, crsTransformerFactory, limitsGenerator, cache, staticTileProviderStore, tileMatrixSetRepository);
         this.providers = providers;
-    }
-
-    @Override
-    public List<String> getConformanceClassUris() {
-        return ImmutableList.of("http://www.opengis.net/spec/ogcapi-tiles-1/0.0/conf/core");
     }
 
     @Override
@@ -104,42 +90,11 @@ public class EndpointTileMultiCollection extends AbstractEndpointTileMultiCollec
     }
 
     @Override
-    public Class<? extends ExtensionConfiguration> getBuildingBlockConfigurationType() {
-        return TilesConfiguration.class;
-    }
-
-    @Override
-    public List<? extends FormatExtension> getFormats() {
-        if (formats==null)
-            formats = extensionRegistry.getExtensionsForType(TileFormatExtension.class);
-        return formats;
-    }
-
-    @Override
     protected ApiEndpointDefinition computeDefinition(OgcApiDataV2 apiData) {
-        ImmutableApiEndpointDefinition.Builder definitionBuilder = new ImmutableApiEndpointDefinition.Builder()
-                .apiEntrypoint("tiles")
-                .sortPriority(ApiEndpointDefinition.SORT_PRIORITY_TILE);
-        final String path = "/tiles/{tileMatrixSetId}/{tileMatrix}/{tileRow}/{tileCol}";
-        final HttpMethods method = HttpMethods.GET;
-        final List<OgcApiPathParameter> pathParameters = getPathParameters(extensionRegistry, apiData, path);
-        final List<OgcApiQueryParameter> queryParameters = getQueryParameters(extensionRegistry, apiData, path);
-        String operationSummary = "fetch a tile with multiple layers, one per collection";
-        Optional<String> operationDescription = Optional.of("The tile in the requested tiling scheme ('{tileMatrixSetId}'), " +
-                "on the requested zoom level ('{tileMatrix}'), with the requested grid coordinates ('{tileRow}', '{tileCol}') is returned. " +
-                "The tile has one layer per collection with all selected features in the bounding box of the tile with the requested properties.");
-        ImmutableOgcApiResourceData.Builder resourceBuilder = new ImmutableOgcApiResourceData.Builder()
-                .path(path)
-                .pathParameters(pathParameters);
-        ApiOperation operation = addOperation(apiData, queryParameters, path, operationSummary, operationDescription, TAGS);
-        if (operation != null)
-            resourceBuilder.putOperations(method.name(), operation);
-        definitionBuilder.putResources(path, resourceBuilder.build());
-
-        return definitionBuilder.build();
+        return null;
     }
 
-    @Path("/{tileMatrixSetId}/{tileMatrix}/{tileRow}/{tileCol}")
+    @Path("/map/tiles/{tileMatrixSetId}/{tileMatrix}/{tileRow}/{tileCol}")
     @GET
     public Response getTile(@Auth Optional<User> optionalUser, @Context OgcApi api,
                             @PathParam("tileMatrixSetId") String tileMatrixSetId, @PathParam("tileMatrix") String tileMatrix,
